test_imaging_confocal: migrate plotting tests
